Add enums and API-agnostic handling for new task and mesh shader stages
* The enums are given after compute, to preserve indices for the normal vertex
  pipeline.
* Mesh dispatches are considered a new action type, rather than being bundled
  into the `Drawcall` type. This will allow them to be distinguished by API
  backends as needed. The UI treats them as drawcalls
* We apply this universally even though it's not relevant to D3D11/GL. It means
  a couple of empty array entries but it should not cause any significant
  issues.
* Shader messages will be identified by group and thread as with compute
  shaders. For mesh shaders there is an additional subdivision to identify them
  by task group, since each task group can submit a grid of mesh groups.

+++ renderdoc/driver/shaders/dxbc/dxbc_container.cpp
@@ -1785,7 +1785,7 @@ DXBCContainer::DXBCContainer(const bytebuf &ByteCode, const rdcstr &debugInfoPat
         desc.regChannelMask = (uint8_t)(el->mask & 0xff);
         desc.channelUsedMask = (uint8_t)(el->rwMask & 0xff);
         desc.regIndex = el->registerNum;
-        desc.semanticIndex = el->semanticIdx;
+        desc.semanticIndex = (uint16_t)el->semanticIdx;
         desc.semanticName = chunkContents + el->nameOffset;
         desc.systemValue = GetSystemValue(el->systemType);
         desc.compCount = (desc.regChannelMask & 0x1 ? 1 : 0) + (desc.regChannelMask & 0x2 ? 1 : 0) +
